refactor: rename CareerPositionDetailsSection to PositionDetails
