--- src/components/career/CareerPositionDetailsSection/index.ts
@@ -1,5 +1,0 @@
-import CareerPositionDetailsSection from "./CareerPositionDetailsSection";
-import { CareerPositionDetailsSectionProps } from "./CareerPositionDetailsSection";
-
-export { CareerPositionDetailsSection };
-export type { CareerPositionDetailsSectionProps };
