+++ src/components/career/PositionDetails/index.ts
@@ -1,0 +1,5 @@
+import { PositionDetails } from "./PositionDetails";
+import { PositionDetailsProps } from "./PositionDetails";
+
+export { PositionDetails };
+export type { PositionDetailsProps };
