+++ src/components/career/index.ts
@@ -1,15 +1,10 @@
 import { CareerGeneralIntro } from "./CareerGeneralIntro";
 import CareerHero from "./CareerHero";
-import { CareerPositionDetailsSection } from "./CareerPositionDetailsSection";
-import type { CareerPositionDetailsSectionProps } from "./CareerPositionDetailsSection";
+import { PositionDetails } from "./PositionDetails";
+import type { PositionDetailsProps } from "./PositionDetails";
 import { CareerPositionList } from "./CareerPositionList";
 import type { CareerPositionListProps } from "./CareerPositionList";
 
-export {
-  CareerGeneralIntro,
-  CareerHero,
-  CareerPositionDetailsSection,
-  CareerPositionList,
-};
+export { CareerGeneralIntro, CareerHero, PositionDetails, CareerPositionList };
 
-export type { CareerPositionDetailsSectionProps, CareerPositionListProps };
+export type { PositionDetailsProps, CareerPositionListProps };
